Created a new ZapModel class to bootstrap mAP logging in all (?) models
