+++ src/zap/__init__.py
@@ -9,10 +9,11 @@ supress_pydantic_warnings()  # noqa
 
 from typing import Any  # noqa
 
-from lightning.pytorch import LightningDataModule  # noqa
+from lightning.pytorch import LightningDataModule, LightningModule  # noqa
 from lightning.pytorch.cli import LightningCLI  # noqa
 from PIL import Image  # noqa
 from torch.utils.data import DataLoader, Dataset  # noqa
+from torchmetrics.detection import MeanAveragePrecision  # noqa
 
 format_lightning_warnings_and_logs()
 
@@ -47,6 +48,32 @@ class Zap():
             return_predictions=True,
             ckpt_path=ckpt_path)
         return preds
+
+
+# TODO: test compatibility with classification and segmentation models
+class ZapModel(LightningModule):
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
+        super().__init__(*args, **kwargs)
+        self.precision = MeanAveragePrecision(class_metrics=True)
+
+    def on_test_epoch_start(self) -> None:
+        self.label_map = self.trainer.datamodule.label_map
+
+    def log_precision(self, precision):
+        for k, v in precision.items():
+            if k == 'classes':  # don't record the classes key; not useful
+                continue
+
+            # handle single class vs multiclass
+            class_values = v.tolist()
+            if not isinstance(class_values, list):
+                if class_values >= 0:  # mAP API returns -1 for missing classes; need to ignore
+                    self.log(k, class_values, on_epoch=True)
+            else:
+                for class_index, val in enumerate(class_values):  # log each class metric separately
+                    if val >= 0:
+                        k = k.replace('_per_class', '_class')
+                        self.log(f'{k}_{self.label_map[class_index]}', val, on_epoch=True)
 
 
 class ZapDataModule(LightningDataModule):
